refactor: rename Period Details Dialog file
File renamed to match component name.
