+++ components/schedule/PeriodDetailsDialog.tsx
